UI_01/Routing/Routing/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
aiokafka==0.13.0
httpx==0.28.1
livekit==1.1.2
livekit-api==1.1.0
numpy==2.2.6
scipy==1.15.3
PyJWT==2.12.1
PyYAML==6.0.3
aiohttp==3.13.3
aiofiles==25.1.0
prometheus_client==0.24.1
requests==2.33.1
google-genai>=1.0.0
faster-whisper>=1.0.0
pytest
pytest-asyncio
asyncpg>=0.30.0
pytz>=2024.1
